crm: Resolve Sass deprecation warnings (#43854)
Mostly straightforward. In a few places I went with `meta.load-css()`
rather than `@use` where the module was being imported only for its CSS
to avoid having to have `as styles1`, `as styles2`, ... `as styles24`.

Also there was one odd bit, if A uses B then A can `@extend` selectors
from B but not vice versa, and there's one place in here that was trying
to do the vice versa. Since it was only being extended in the one place,
from two places that declared identical styles, I just copied it into
the extended location. This moves the rule around in the built file, but
AFAICT it won't make any difference to the styling.