ci: Further fix tests on plugin release branches (#44003)
PR #43953 added a `composer require` to add missing packages, but it
breaks on older PHP versions because the composer.lock has too-new of
packages locked. Use `--ignore-platform-reqs` to avoid that problem;
later code will handle doing the needed `composer update`.

We also need to add the missing packages in the Phan job, which was
missed in the earlier PR.