eslint: Lint json files (#44009)
Use the relatively new `@eslint/json` plugin to lint json files.

This unfortunately needs some changes to our eslint configurations.
Many existing rules expect a sourceCode object with an interface that's
compatible with the standard SourceCode, and the devs of `@eslint/json`
explicitly refuse to do that. Instead they say that every configuration
object should specify `files`.

On the plus side, many of the configs don't actually need editing since
eslint seems fine with turning _off_ rules even if they're not loaded
for a file, it's only when they're enabled that we need the `files`.

As for actual linting, turns out most of our JSON files in the monorepo
are already pretty good. One file had a duplicate key in an object,
prettier had a bunch of reformatting to do, and we're ignoring the i18n
json files in `projects/packages/jetpack-mu-wpcom/language/`.