storybook: Delete broken stories (#44073)
Back in #37124 we renamed 33 broken Storybook stories, with the idea
that people could rename them back if they wanted to fix them. In the
year+ since, 6 have been removed because the parent component was
removed.

This removes the other 27, since it seems likely no one is going to get
around to fixing them any time soon so there's little point in having
them come up in refactors and linting. If people want to fix them
someday (instead of writing a working story from scratch), they can pull
them out of the git history.