Admin Menu: Show "Jetpack > Stats" to Calypso users (#44090)
Displays the "Jetpack > Stats" menu to all users (previously, only visible to sites with the classic admin interface active).

On a follow-up, we'll start showing a callout in the existing Stats top-level menu, informing users that we moved the screen to Jetpack > Stats. In the meantime, both menus (Stats, and Jetpack > Stats) will coexist for sites with the default admin interface.