Clean up unneeded React imports (#44136)
React 17 made it no longer necessary to do `import React from 'react'`
all over the place, and in fact now discourages that. Let's clean it up.

The [codemod] did the first part of the work, but a bunch of TypeScript
and jsdoc stuff needed manual fixing after.

This also adjusts the `jsx` option in jsconfig.json files to the correct setting now that we're using the modern transform.

[codemod]: https://legacy.reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html#removing-unused-react-imports